Rename "recipes" package to "instances"
